Add Rating weighting in Data Wrangling section.
Added a new column "Weighted_rating" to our dataset representing a
weighted average of movie ratings in order to take into account the
number of people who gave a rating for each movie.

Modified the rest of the code to do the analysis based on this new
column rather then the original "Rating" column.